remove government information form
