+++ apps/web/src/pages/Applications/ApplicationProfilePage/ApplicationProfilePage.tsx
@@ -17,7 +17,6 @@ import ProfileFormProvider from "~/components/Profile/components/ProfileFormCont
 import PersonalInformation from "~/components/Profile/components/PersonalInformation/PersonalInformation";
 import WorkPreferences from "~/components/Profile/components/WorkPreferences/WorkPreferences";
 import DiversityEquityInclusion from "~/components/Profile/components/DiversityEquityInclusion/DiversityEquityInclusion";
-import GovernmentInformation from "~/components/Profile/components/GovernmentInformation/GovernmentInformation";
 import LanguageProfile from "~/components/Profile/components/LanguageProfile/LanguageProfile";
 import poolCandidateMessages from "~/messages/poolCandidateMessages";
 import ContactEmailCard from "~/components/ContactEmailCard/ContactEmailCard";
@@ -167,7 +166,6 @@ export const ApplicationProfile = ({ application }: ApplicationPageProps) => {
           <DiversityEquityInclusion {...sectionProps} />
         </div>
         <CitizenVeteranPriority {...sectionProps} />
-        <GovernmentInformation query={application.user} />
         <LanguageProfile
           {...sectionProps}
           application={{
